Update worker to not System.exit
In JDK 21 you cannot intercept System.exit, so we stopped doing that in
rules_scala_annex. Instead we throw an error that we catch in the base
class of the worker and handle instead of exiting from the worker.

This updates the rules to adopt that new behavior instead of
System.exit, which would now just crash the worker.